+++ lib/arachni/browser.rb
@@ -896,11 +896,26 @@ class Browser
 
             begin
                 input.set( value.to_s )
-                # Disabled inputs and such...
+            # Disabled inputs and such...
             rescue Watir::Exception::ObjectDisabledException,
                 Watir::Exception::ObjectReadOnlyException,
                 Selenium::WebDriver::Error::InvalidElementStateError => e
                 print_debug_level_2 "Could not fill in form input '#{name_or_id}'" <<
+                                        " because: #{e} [#{e.class}"
+            end
+        end
+
+        form.selects.each do |input|
+            name_or_id = name_or_id_for( input )
+            value      = inputs ? inputs[name_or_id] : value_for( input )
+
+            begin
+                form.select_list( name: name_or_id ).select_value( value.to_s )
+            # Disabled inputs and such...
+            rescue Watir::Exception::ObjectDisabledException,
+                Watir::Exception::ObjectReadOnlyException,
+                Selenium::WebDriver::Error::InvalidElementStateError => e
+                print_debug_level_2 "Could not fill in form select '#{name_or_id}'" <<
                                         " because: #{e} [#{e.class}"
             end
         end