+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## _Under development_
 
+- `Browser` -- Fixed issue causing `select` inputs in forms to not be set.
 - `Issue`
     - `#recheck` -- Rechecks the existence of the issue.
 - `Framework` -- Split into `Parts`: